+++ Android/Sunshine/S10.04-Exercise-Notifications/app/src/main/java/com/example/android/sunshine/data/SunshinePreferences.java
@@ -18,6 +18,7 @@ package com.example.android.sunshine.data;
 import android.content.Context;
 import android.content.SharedPreferences;
 import android.preference.PreferenceManager;
+import android.support.v4.content.ContextCompat;
 
 import com.example.android.sunshine.R;
 
@@ -153,6 +154,16 @@ public final class SunshinePreferences {
         }
 
         return spContainBothLatitudeAndLongitude;
+    }
+
+    public static boolean areNotificationsEnabled(Context context) {
+        String displayNotificationsKey = context.getString(R.string.pref_enable_notifications_key);
+
+        boolean shouldDisplayNotificationsByDefault = context
+                .getResources().getBoolean(R.bool.show_notifications_by_default);
+
+        SharedPreferences sp = PreferenceManager.getDefaultSharedPreferences(context);
+        return sp.getBoolean(displayNotificationsKey, shouldDisplayNotificationsByDefault);
     }
 
     /**